backend/requirements.txt
# FastAPI and server
fastapi>=0.104.0,<0.115.0
uvicorn[standard]>=0.24.0,<0.30.0
gunicorn>=21.0.0,<23.0.0

# Database
sqlalchemy>=2.0.0,<3.0.0
alembic>=1.12.0,<2.0.0
psycopg2-binary>=2.9.0,<3.0.0
asyncpg>=0.29.0,<0.30.0

# Data validation
pydantic[email]>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0

# Authentication
python-jose[cryptography]>=3.3.0,<4.0.0
python-multipart>=0.0.6,<0.1.0
passlib[bcrypt]>=1.7.4,<2.0.0

# HTTP client
httpx>=0.25.0,<0.28.0

# Task queue
celery>=5.3.0,<6.0.0
redis>=5.0.0,<6.0.0

# AI/ML
google-generativeai>=0.3.0,<0.9.0
openai>=1.0.0,<2.0.0

# Video processing
ffmpeg-python>=0.2.0,<0.3.0

# Security
cryptography>=41.0.0,<43.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0

# Supabase
supabase>=2.0.0,<3.0.0

# Testing
pytest>=7.0.0,<9.0.0
pytest-asyncio>=0.21.0,<0.25.0
pytest-cov>=4.0.0,<6.0.0

# Code quality
black>=23.0.0
ruff>=0.1.0
mypy>=1.0.0
